Module 2 Assginment
Write a program that sorts Strings. The program should prompt the user for the number of Strings to read, and then read the Strings from standard input one at a time. Then it should sort the Strings into alphabetical order using any sorting algorithm with which you are familiar, and print them to the console. The commentary material on console I/O, arrays, and the String.compareTo() method should all be useful, so be sure to read the commentary! Do not use any classes in the java.util package; specifically, do not use java.util.Arrays or java.util.Vector.