Module 4 Assignment
Write a cash register class. It should have the following methods:


AddTransaction(amount) – add an amount to a running total and increments the number of transactions
TransactionCount( ) – returns the number of transactions received
Total( ) – returns the sum of the amounts of the transactions
ResetTransactions( ) – sets the total amount and transaction count to 0
Register Count( ) – returns the number of cash registers created
Next, write an application that instantiates at least two cash register objects to prove that your class works.