Module 5 Assignment
Part 1:
For each of the following lines of code, tell me whether the assignment will compile, and if so, if it will succeed or fail at runtime. For any that fail to compile, pretend that the right-hand side of that statement is null, and answer the remaining questions assuming this change has been made. 

Part 2:
Add the following declaration to interface Vehicle:

int getWeightInPounds( );


Then add the minumum amount of additional code to make all the classes that implement Vehicle compile again. Assume that a small car weighs 1,000 pounds, and a large pickup truck weighs 2,500 pounds. Try to write as little redundant code as possible. You may introduce new classes if this would help meet this goal.

Write a class ParkingGarage which simulates a garage for parking Vehicles. The garage should have have enough capacity to park 20 Vehicles or 25,000 pounds. It should refuse to park any Vehicle that would push the garage over either of these limits. Note that ParkingGarage should be a well-behaved class; in particular, it should be possible to create multiple independent ParkingGarages in one program.

ParkingGarage should have a main routine which demonstrates its capabilities. main( ) need not accept any user input.