Module 6 Assignment
Use the collections framework and generic types to create a bowling score tracker collection. Design a data structure that will be used as elements in a collection. Each element should contain a unique name of the bowler and a list of scores and dates of the games they have played.

Then print to the screen each:

Bowler name
Number of games
Average score of all of the games
Score of their last game
Date of their last game
Provide test cases to prove your collection works