Add files via upload
Write a program that will display lines of a text you specify. It should take two arguments: a file or a URL and the number of lines to display. If the number of lines given is positive, it should display the first n lines of the text.

If it is 0, it should display all the lines of the text. If it is a negative number, it should display the last n lines of the text. Your program should handle errors appropriately.